+++ src/com/page/node/events/events.js
@@ -1,17 +1,15 @@
 import {h, Component}					from 'preact/preact';
 import PageNavRoot						from '../../nav/root';
 
-//import ContentGroup						from 'com/content-group/group';
-import ContentEvents from 'com/content-events/events';
+import ContentGroup						from 'com/content-group/group';
 
 export default class PageEvents extends Component {
     render( props ) {
         return (
 			<div id="content">
 				<PageNavRoot {...props} />
-        <ContentEvents {...props} />
+        <ContentGroup {...props} />
 			</div>
         );
-      //<ContentGroup {...props} />
     }
 }
